# [ingest] use BulkRequest with 200 docs per chunk to ingest results faster