## Summary

Improve ingestion speed with `BulkRequest`

Before:
```
15:24:47.672 [INFO ] ES_Client - Ingesting to stats cluster: 6136 docs
15:26:06.109 [INFO ] ES_Client - Ingestion is completed
// 1 min 19 sec
...
15:26:11.165 [INFO ] ES_Client - Ingesting to stats cluster: 2662 docs
15:26:46.768 [INFO ] ES_Client - Ingestion is completed
// 35 sec
```

After
```
15:35:21.947 [INFO ] ES_Client - Ingesting to stats cluster: 6136 docs
15:35:37.739 [INFO ] ES_Client - Ingestion is completed
// 16 sec
...
15:35:37.765 [INFO ] ES_Client - Ingesting to stats cluster: 281 docs
15:35:37.913 [INFO ] ES_Client - Ingestion is completed
// 0.15 sec
```

### Checklist

Delete any items that are not applicable to this PR.

- [x] Branch is successfully tested on [Kibana CI](https://kibana-ci.elastic.co/job/elastic+kibana+load-testing/)
- [ ] Unit tests are added